Move argument class to its own header
