+++ components/core/wf/code_generation/function_description.cc
@@ -5,31 +5,6 @@
 
 namespace wf {
 
-argument::argument(const std::string_view name, type_variant type,
-                   const argument_direction direction, const std::size_t index)
-    : impl_(std::make_shared<const impl>(
-          impl{std::string(name), std::move(type), direction, index})) {}
-
-std::variant<scalar_expr, matrix_expr, compound_expr> argument::create_symbolic_input() const {
-  return std::visit(
-      [&](const auto& type_concrete) -> std::variant<scalar_expr, matrix_expr, compound_expr> {
-        return detail::create_function_input(type_concrete, index());
-      },
-      type());
-}
-
-std::size_t hash_struct<argument>::operator()(const argument& arg) const noexcept {
-  std::size_t hash = hash_string_fnv(arg.name());
-  hash = hash_args(hash, arg.type());
-  hash = hash_combine(hash, static_cast<std::size_t>(arg.direction()));
-  return hash_combine(hash, arg.index());
-}
-
-bool is_identical_struct<argument>::operator()(const argument& a, const argument& b) const {
-  return a.name() == b.name() && are_identical(a.type(), b.type()) &&
-         are_identical(a.direction(), b.direction()) && are_identical(a.index(), b.index());
-}
-
 function_description::function_description(std::string name) noexcept
     : impl_(std::make_shared<impl>(std::move(name))) {}
